requirements.txt
pytest
discord.py
python-dotenv
requests
pytz
watchdog
